fix(ai): sanitize two new tracing:: log sites and split embedding out of documents
documents/mod.rs sat at exactly R8's 1400-LOC hard cap with zero headroom,
bought partly by trimming comments; the next one-line change would fail the
architecture test. Extracts the self-contained embedding round-trip +
posting-vector cache resolver into documents/embedding.rs (re-exported flat
at documents:: so every existing crate::documents::X call site is
unaffected), leaving 185 lines of real headroom.

Also sanitizes two tracing::warn! sites this branch introduced in the same
area (embedding_config's rusqlite::Error read failure, and embed()'s
provider-call failure) -- check-log-error-leaks.mjs only scans log:: macros,
so tracing:: is a blind spot the guard can't see; one of the two wraps a raw
reqwest::Error, and Gemini authenticates via a `?key=` query param.
